# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pptx2pdf",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/graph-credentials": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Graph Credentials",
        "operationId": "get_graph_credentials_api_admin_graph_credentials_get",
        "parameters": [
          {
            "name": "pptx2pdf_admin",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pptx2Pdf Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphCredentialsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Put Graph Credentials",
        "description": "先测后存：五步自检全绿才写库。\n\n配错的凭证一旦入库，三期的引擎每次转换都会去读它——期间所有走\nGraph 的任务都会失败，而失败原因（一个 Graph 原始报错）与「你配错\n了」之间隔着数层抽象。这条顺序把那个窗口彻底关闭：run_selftest 在\nsave_credentials 之前调用且不捕获它的异常，任何一步——包括\nrun_selftest 内部未预期的穿透异常——都会在触达 save_credentials 之前\n终止请求，数据库不会被写入半条记录。",
        "operationId": "put_graph_credentials_api_admin_graph_credentials_put",
        "parameters": [
          {
            "name": "pptx2pdf_admin",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pptx2Pdf Admin"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphCredentialsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelftestResultDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GraphCredentialsDto": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "drive_path": {
            "type": "string",
            "title": "Drive Path"
          },
          "secret_configured": {
            "type": "boolean",
            "title": "Secret Configured"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "client_id",
          "site_id",
          "drive_path",
          "secret_configured"
        ],
        "title": "GraphCredentialsDto",
        "description": "读配置的响应。client_secret 绝不出现在这里——解密回显等于把凭证\n明文发到浏览器，那么加密存库本身就失去意义。"
      },
      "GraphCredentialsUpdate": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Tenant Id"
          },
          "client_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "maxLength": 512,
            "title": "Client Secret",
            "default": ""
          },
          "site_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Site Id"
          },
          "drive_path": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Drive Path",
            "default": "pptx2pdf-staging"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "client_id",
          "site_id"
        ],
        "title": "GraphCredentialsUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SelftestResultDto": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/SelftestStepDto"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "steps"
        ],
        "title": "SelftestResultDto"
      },
      "SelftestStepDto": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "step",
          "ok",
          "detail"
        ],
        "title": "SelftestStepDto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}